tweaking

+++ pablo/lib/animations.py
@@ -68,8 +68,8 @@ class RunnerAnimation(Animation):
 	def getFrame(self, frame):
 		buffer = []
 
-		headLocation = int((frame * self.config["speed"]["step"])) % self.config["pixels"]
-		tailLocation = int((headLocation - self.config["tail"])) % self.config["pixels"]
+		headLocation = (frame * self.config["speed"]["step"]) % self.config["pixels"]
+		tailLocation = (headLocation - self.config["tail"]) % self.config["pixels"]
 		print "Head: %d, Tail: %d" % (headLocation, tailLocation)
 		for i in range(self.config["pixels"]):
 			if i > tailLocation and i <= headLocation:
